scripts/requirements_ml.txt
numpy>=1.19.2
pandas>=1.2.4
scikit-learn>=0.24.2
tensorflow>=2.6.0
psycopg2-binary>=2.9.1
